Support direct kurt plugin import and output invalid arguments.
The -k option now can also take a python import path, rather than a filename to
import.

Files and directories listed at the end of the command line will be listed as
invalid if the file is not of a supported extension, or the directory tree does
not contain at least one file of a supported extension.